Zmiana typu danych na decimal

+++ WindowsFormsApplication1/FBGManagement/Grating.cs
@@ -8,41 +8,41 @@ namespace WindowsFormsApplication1.FBGManagement
 {
     class Grating
     {
-        public double length { get; }
-        public double period { get; }
-        public double refractiveIndexModulation { get; }
-        public double neff { get; }
-        public double lambdaB {get;}
+        public decimal length { get; }
+        public decimal period { get; }
+        public decimal refractiveIndexModulation { get; }
+        public decimal neff { get; }
+        public decimal lambdaB {get;}
         //period=lambdaB/(2*pi*neff);
         public int parts { get; }
 
-        public Grating(double period, double length, double refractiveIndexModulation, double neff, int parts)
+        public Grating(decimal period, decimal length, decimal refractiveIndexModulation, decimal neff, int parts)
         {
             this.neff = neff;
             this.length = length;
             this.period = period;
             this.refractiveIndexModulation = refractiveIndexModulation;
             this.parts = parts;
-            lambdaB = period / (2 * Math.PI * neff);
+            lambdaB = period / (2 * (decimal)Math.PI * neff);
         }
-        public Grating(double period, double length, double refractiveIndexModulation, double neff) :  this(period, length,refractiveIndexModulation,neff, 10)
+        public Grating(decimal period, decimal length, decimal refractiveIndexModulation, decimal neff) :  this(period, length,refractiveIndexModulation,neff, 10)
         {
 
         }
-        public Grating(double period, double length, double refractiveIndexModulation) : this(period, length, refractiveIndexModulation, 1.44688)
+        public Grating(decimal period, decimal length, decimal refractiveIndexModulation) : this(period, length, refractiveIndexModulation, 1.44688m)
         {
 
         }
-        public Grating(double period, double length) : this(period, length, 0.0001)
+        public Grating(decimal period, decimal length) : this(period, length, 0.0001m)
         {
         }
-        public Grating() : this(0.0000015312 / (2 * Math.PI * 1.44688), 0.1)
+        public Grating() : this(0.0000015312m / (2m * (decimal)Math.PI * 1.44688m), 0.1m)
         {
-            this.neff = 1.44688;
-            length = 0.1; //metra
-            lambdaB = 0.0000015312;
-            period = 0.0000015312 / (2 * Math.PI * neff);
-            refractiveIndexModulation = 0.0004;
+            this.neff = 1.44688m;
+            length = 0.1m; //metra
+            lambdaB = 0.0000015312m;
+            period = 0.0000015312m / (2m * (decimal)Math.PI * neff);
+            refractiveIndexModulation = 0.0004m;
         }
     }
 }
